Ajustando estilos e pegando as imagens do carrosel diretamente de um arquivo JSON
