--- src/components/Button/Botao.scss
@@ -1,7 +1,0 @@
-@media(max-width: 500px){
-    a{
-        padding: 0px 0px;
-        background-color: #374151;
-        border: none;
-    }
-}
