+++ src/components/BuyBox/BuyBox.scss
@@ -136,11 +136,16 @@
 .buy {
     width: 13.75rem;
     height: 3rem;
-    background-color: #FFB31F;
+    background-color: #C92071;
     font-weight: 700;
     font-family: "Inter", sans-serif;
     color: #fff;
     border: none;
     border-radius: 4px;
     margin-top: 1.5rem;
+}
+
+.buy:hover{
+    background-color: #8F8F8F;
+    color: #C92071;
 }
