+++ src/components/Gallery/Gallery-slide.scss
