--- src/components/Gallery/gallery-slide.jsx
@@ -1,40 +1,0 @@
-import 'bootstrap/dist/css/bootstrap.min.css';
-import "./gallery-slide.scss";
-
-const Gallery = () => {
-
-
-    return (
-
-      <div id="carouselExampleRide" className="carousel slide" data-bs-ride="true">
-  <div className="carousel-inner">
-    <div className="carousel-item active">
-      <img src="images/home-slide-6.jpeg" className="d-block w-100" alt="..."/>
-    </div>
-    <div className="carousel-item">
-      <img src="images/home-slide-2.jpeg" className="d-block w-100" alt="..."/>
-    </div>
-    <div className="carousel-item">
-      <img src="images/home-slide-3.jpeg" className="d-block w-100" alt="..."/>
-    </div>
-    <div className="carousel-item">
-      <img src="images/home-slide-4.jpeg" className="d-block w-100" alt="..."/>
-    </div>
-    <div className="carousel-item">
-      <img src="images/home-slide-5.jpeg" className="d-block w-100" alt="..."/>
-    </div>
-  </div>
-  <button className="carousel-control-prev" type="button" data-bs-target="#carouselExampleRide" data-bs-slide="prev">
-    <span className="carousel-control-prev-icon" aria-hidden="true"></span>
-    <span className="visually-hidden">Previous</span>
-  </button>
-  <button className="carousel-control-next" type="button" data-bs-target="#carouselExampleRide" data-bs-slide="next">
-    <span className="carousel-control-next-icon" aria-hidden="true"></span>
-    <span className="visually-hidden">Next</span>
-  </button>
-</div>
-
-  )
-}
-
-export default Gallery
