+++ src/components/Header/Header.scss
@@ -13,7 +13,7 @@
 
   .logo-header {
     margin-left: 20px;
-    padding-left: 12px;
+    padding-left: 11px;
   }
 
   #busca {
