+++ src/json/gallery.json
@@ -1,0 +1,17 @@
+[
+    {
+        "image" : "images/home-slide-6.jpeg"
+    },
+    {
+        "image" : "images/home-slide-2.jpeg"
+    },
+    {
+        "image" : "images/home-slide-3.jpeg"
+    },
+    {
+        "image" : "images/home-slide-4.jpeg"
+    },
+    {
+        "image" : "images/home-slide-5.jpeg"
+    }
+]
