+++ src/json/produtosListing.json
@@ -1,7 +1,7 @@
 [
     {
         "id": 1,
-        "name": "Tênis Nike Air Force 1",
+        "name": "Air Force 1",
         "image": [
             "/images/details/tenis1/lado.png",
             "/images/details/tenis1/ladoDireito.png",
@@ -121,7 +121,7 @@
     },
     {
         "id": 1,
-        "name": "Tênis Nike Air Force 1",
+        "name": "Air Force 1",
         "image": [
             "/images/details/tenis1/lado.png",
             "/images/details/tenis1/ladoDireito.png",
@@ -238,6 +238,67 @@
         "description": "O Air Naldo é uma edição especial da linha de calçados do Superstar Naldo Benny, projetada em colaboração com 'The Shop', uma marca que conecta comunidades. Este modelo traz um novo olhar para a linha de sapatos assinados pelo Rei, oferecendo uma sensação fresca e acima do padrão após um corte limpo.",
         "sizeOptions": [37,39, 41, 42],
         "colorOptions": ["#6FEEFF", "#C92071", "#5E5E5E", "#6D70B7"]
+    },
+    {
+        "id": 2,
+        "name": "Lebron XXI",
+        "image": [
+            "/images/details/tenis2/lado.png",
+            "/images/details/tenis2/ladoDireito.png",
+            "/images/details/tenis2/lateral.png",
+            "/images/details/tenis2/lado.png",
+            "/images/details/tenis2/sola.png"
+        ],
+        "price": 300,
+        "priceDiscount": 100,
+        "ref": "Casual | Nike | REF:38416711",
+        "rating": 4.5,
+        "reviews": "(90 avaliações)",
+        "type": "Tênis",
+        "description": "O Nike LeBron XXI x The Shop é uma edição especial da linha de calçados de basquete do LeBron James, projetada em colaboração com 'The Shop', uma marca que conecta comunidades e série de conversas. Este modelo traz um novo olhar para a linha de sapatos assinados pelo Rei, oferecendo uma sensação fresca e acima do padrão após um corte limpo.",
+        "sizeOptions": [39, 41, 42],
+        "colorOptions": ["#6FEEFF", "#C92071", "#5E5E5E", "#6D70B7"]
+    },
+    {
+        "id": 4,
+        "name": "Air Jordan 37 Low",
+        "image": [
+            "/images/details/tenis4/lado.png",
+            "/images/details/tenis4/ladoDireito.png",
+            "/images/details/tenis4/lateral.png",
+            "/images/details/tenis4/costas.png",
+            "/images/details/tenis4/sola.png"
+        ],
+        "price": 499,
+        "priceDiscount": 100,
+        "ref": "Casual | Nike | REF:38416711",
+        "rating": 4.5,
+        "reviews": "(95 avaliações)",
+        "type": "Tênis",
+        "description": "A Air Jordan é uma icônica linha de calçados de basquete produzida pela Nike, Inc., criada para o ex-jogador de basquete Michael Jordan durante seu tempo com o Chicago Bulls. Lançadas em 1985, as Air Jordan revolucionaram o jogo com seu carisma arrojado e se tornaram um ícone cultural, influenciando tanto o esporte quanto a moda urbana.",
+        "sizeOptions": [39, 40, 41, 42, 43],
+        "colorOptions": ["#6FEEFF", "#C92071", "#5E5E5E", "#6D70B7"]
+    },
+    {
+        "id": 6,
+        "name": "Air Naldo",
+        "image": [
+            "/images/produc-image-1.png",
+            "/images/produc-image-1.png",
+            "/images/produc-image-1.png",
+            "/images/produc-image-1.png",
+            "/images/produc-image-1.png"
+        ],
+        "price": 700,
+        "priceDiscount": 0,
+        "ref": "Casual | Nike | REF:38416711",
+        "rating": 4.9,
+        "reviews": "(99 avaliações)",
+        "type": "Tênis",
+        "description": "O Air Naldo é uma edição especial da linha de calçados do Superstar Naldo Benny, projetada em colaboração com 'The Shop', uma marca que conecta comunidades. Este modelo traz um novo olhar para a linha de sapatos assinados pelo Rei, oferecendo uma sensação fresca e acima do padrão após um corte limpo.",
+        "sizeOptions": [37,39, 41, 42],
+        "colorOptions": ["#6FEEFF", "#C92071", "#5E5E5E", "#6D70B7"]
     }
+    
   
     ]
