+++ src/pages/HomePage/Home.jsx
@@ -1,7 +1,7 @@
 
 import ProductListing from "../../components/ProductListing/ProductListing";
 import './Home.scss'
-import Gallery from "../../components/Gallery/gallery-slide";
+import Gallery from "../../components/Gallery/Gallery-slide";
 import Collection from "../../components/Collection/Collection";
 import Layout from "../Layout/Layout";
 import produtosHome from '../../json/produtosListing.json'
